+++ src/bgg/lut/mod.rs
@@ -1,2 +1,1 @@
 pub mod public_lut;
-pub mod utils;
