+++ src/bgg/mod.rs
@@ -5,6 +5,7 @@ pub mod lut;
 pub mod norm_simulator;
 pub mod public_key;
 pub mod sampler;
+pub mod utils;
 
 pub use digits_to_int::DigitsToInt;
 pub use encoding::BggEncoding;
